requirements.txt
flask[async]==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio==3.4.3
requests